+++ conventions/src/main/kotlin/otel.javaagent-bootstrap.gradle.kts
@@ -11,9 +11,9 @@ base.archivesName.set(projectDir.parentFile.name + "-bootstrap")
 dependencies {
   compileOnly("io.opentelemetry:opentelemetry-api")
   compileOnly(project(":instrumentation-api"))
-  compileOnly(project(":javaagent-instrumentation-api"))
+  compileOnly(project(":javaagent-bootstrap"))
 
   testImplementation("io.opentelemetry:opentelemetry-api")
   testImplementation(project(":instrumentation-api"))
-  testImplementation(project(":javaagent-instrumentation-api"))
+  testImplementation(project(":javaagent-bootstrap"))
 }
