+++ instrumentation/aws-lambda/aws-lambda-events-2.2/javaagent/build.gradle.kts
@@ -13,6 +13,8 @@ muzzle {
 }
 
 dependencies {
+  compileOnly(project(":javaagent-bootstrap"))
+
   implementation(project(":instrumentation:aws-lambda:aws-lambda-core-1.0:library"))
 
   implementation(project(":instrumentation:aws-lambda:aws-lambda-events-2.2:library")) {
