+++ instrumentation/jboss-logmanager-1.1/javaagent/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
   library("org.jboss.logmanager:jboss-logmanager:1.1.0.GA")
 
   compileOnly(project(":instrumentation-appender-api-internal"))
+  compileOnly(project(":javaagent-bootstrap"))
 
   // ensure no cross interference
   testInstrumentation(project(":instrumentation:java-util-logging:javaagent"))
