+++ instrumentation/logback/logback-appender-1.0/javaagent/build.gradle.kts
@@ -18,6 +18,7 @@ dependencies {
   library("ch.qos.logback:logback-classic:0.9.16")
 
   compileOnly(project(":instrumentation-appender-api-internal"))
+  compileOnly(project(":javaagent-bootstrap"))
 
   implementation(project(":instrumentation:logback:logback-appender-1.0:library"))
