+++ javaagent-bootstrap/build.gradle.kts
@@ -7,6 +7,7 @@ group = "io.opentelemetry.javaagent"
 
 dependencies {
   implementation(project(":instrumentation-api"))
+  implementation(project(":instrumentation-appender-api-internal"))
   implementation("org.slf4j:slf4j-api")
   implementation("org.slf4j:slf4j-simple")
