+++ javaagent-bootstrap/src/main/java/io/opentelemetry/javaagent/bootstrap/AgentLogEmitterProvider.java
@@ -3,15 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.api.appender.internal;
+package io.opentelemetry.javaagent.bootstrap;
 
 import io.opentelemetry.instrumentation.api.appender.internal.LogEmitterProvider;
 import io.opentelemetry.instrumentation.api.appender.internal.LogEmitterProviderHolder;
 
-/**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at
- * any time.
- */
 public final class AgentLogEmitterProvider {
 
   private static final LogEmitterProviderHolder delegate = new LogEmitterProviderHolder();
