+++ javaagent-tooling/src/test/groovy/io/opentelemetry/javaagent/tooling/OpenTelemetryInstallerTest.groovy
@@ -8,7 +8,7 @@ package io.opentelemetry.javaagent.tooling
 import io.opentelemetry.api.GlobalOpenTelemetry
 import io.opentelemetry.extension.noopapi.NoopOpenTelemetry
 import io.opentelemetry.instrumentation.api.config.Config
-import io.opentelemetry.javaagent.instrumentation.api.appender.internal.AgentLogEmitterProvider
+import io.opentelemetry.javaagent.bootstrap.AgentLogEmitterProvider
 import spock.lang.Specification
 
 class OpenTelemetryInstallerTest extends Specification {
